refactor(dos): Move `elapsedTime()` from the client to the scanner
This utility function is actually used exclusively by the scanner, not
the client, so move it to the former.
