+++ plugins/scanners/dos/src/main/kotlin/DOSUtils.kt
@@ -6,6 +6,9 @@
 
 package org.ossreviewtoolkit.plugins.scanners.dos
 
+import java.time.Duration
+import java.time.Instant
+
 import org.ossreviewtoolkit.model.Package
 
 internal fun Collection<Package>.getDosPurls(): List<String> =
@@ -14,3 +17,16 @@ internal fun Collection<Package>.getDosPurls(): List<String> =
             // Encode a path within the source code to the PURL.
             ?: "${pkg.purl}#${pkg.vcsProcessed.path}"
     }
+
+/**
+ * Elapsed time for a scanjob.
+ */
+internal fun elapsedTime(startTime: Instant): String {
+    val currentTime = Instant.now()
+    val duration = Duration.between(startTime, currentTime)
+    val hours = duration.toHours()
+    val minutes = duration.toMinutesPart()
+    val seconds = duration.toSecondsPart()
+
+    return String.format("%02d:%02d:%02d", hours, minutes, seconds)
+}
